refactor(api): move hooks from api files to separate hook files for better organization and maintainability chore(api): remove unused imports and code related to hooks in api files to clean up the codebase feat(hooks): implement custom hooks for arranger, category, exhibition, material, and season to encapsulate fetching logic and state management
