+++ client/src/api/arranger.ts
@@ -2,8 +2,6 @@ import { type paths, type components } from "../types/api";
 import { endpoint } from "./util";
 import { useState, useEffect } from "react";
 
-type Arranger = components["schemas"]["Arranger"];
-
 export async function getArranger(
   arrangerId: number
 ): Promise<
@@ -15,25 +13,4 @@ export async function getArranger(
     throw new Error(`Failed to fetch arranger with ID ${arrangerId}`);
   }
   return response.json();
-}
-
-export function useArranger(arrangerId?: number): Arranger | null {
-  const [arranger, setArranger] = useState<Arranger | null>(null);
-  useEffect(() => {
-    async function fetchArranger(arrangerId: number) {
-      try {
-        const fetchedArranger = await getArranger(arrangerId);
-        setArranger(fetchedArranger);
-      } catch (error) {
-        console.error(`Failed to fetch arranger ${arrangerId}:`, error);
-        setArranger(null); // エラー時は null を設定
-      }
-    }
-    if (arrangerId === undefined) {
-      setArranger(null); // arrangerIdが未定義の場合は null を設定
-      return;
-    }
-    fetchArranger(arrangerId);
-  }, [arrangerId]);
-  return arranger;
 }
