+++ client/src/api/season.ts
@@ -1,8 +1,5 @@
 import { type paths, type components } from "../types/api";
 import { endpoint } from "./util";
-import { useState, useEffect } from "react";
-
-type Season = components["schemas"]["Season"];
 
 export async function getSeason(
   seasonId: number
@@ -13,25 +10,4 @@ export async function getSeason(
     throw new Error(`Failed to fetch season with ID ${seasonId}`);
   }
   return response.json();
-}
-
-export function useSeason(seasonId?: number): Season | null {
-  const [season, setSeason] = useState<Season | null>(null);
-  useEffect(() => {
-    async function fetchSeason(seasonId: number) {
-      try {
-        const fetchedSeason = await getSeason(seasonId);
-        setSeason(fetchedSeason);
-      } catch (error) {
-        console.error(`Failed to fetch season ${seasonId}:`, error);
-        setSeason(null); // エラー時は null を設定
-      }
-    }
-    if (seasonId === undefined) {
-      setSeason(null); // seasonIdが未定義の場合は null を設定
-      return;
-    }
-    fetchSeason(seasonId);
-  }, [seasonId]);
-  return season;
 }
